+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D11/DX11d3d.cpp
@@ -34,12 +34,12 @@ namespace enigma {
 namespace enigma_user
 {
 
-void d3d_depth_clear() {
+void d3d_clear_depth() {
 	// Clear the depth buffer.
 	m_deviceContext->ClearDepthStencilView(m_depthStencilView, D3D11_CLEAR_DEPTH, 1.0f, 0);
 }
 
-void d3d_depth_clear_value(float value) {
+void d3d_clear_depth_value(float value) {
 	// Clear the depth buffer.
 	m_deviceContext->ClearDepthStencilView(m_depthStencilView, D3D11_CLEAR_DEPTH, value, 0);
 }
