+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D9/DX9d3d.cpp
@@ -54,11 +54,11 @@ D3DFOGMODE fogmodes[3] = {
 namespace enigma_user
 {
 
-void d3d_depth_clear() {
-  d3d_depth_clear_value(1.0f);
+void d3d_clear_depth() {
+  d3d_clear_depth_value(1.0f);
 }
 
-void d3d_depth_clear_value(float value) {
+void d3d_clear_depth_value(float value) {
 	d3dmgr->Clear(0, NULL, D3DCLEAR_ZBUFFER, D3DCOLOR_XRGB(0, 0, 0), value, 0);
 }
