+++ ENIGMAsystem/SHELL/Graphics_Systems/Direct3D9/DX9screen.cpp
@@ -78,158 +78,6 @@ namespace enigma
 	unsigned gui_height;
 }
 
-static inline void draw_back()
-{
-  using enigma_user::background_x;
-  using enigma_user::background_y;
-  using enigma_user::background_visible;
-  using enigma_user::background_alpha;
-  using enigma_user::background_xscale;
-  using enigma_user::background_yscale;
-  using enigma_user::background_htiled;
-  using enigma_user::background_vtiled;
-  using enigma_user::background_hspeed;
-  using enigma_user::background_vspeed;
-  using enigma_user::background_index;
-  using enigma_user::background_coloring;
-  using enigma_user::draw_background_tiled_ext;
-  using enigma_user::draw_background_ext;
-  // Draw the rooms backgrounds
-  for (int back_current = 0; back_current < 8; back_current++) {
-    if (background_visible[back_current] == 1) {
-      if (enigma_user::background_exists(background_index[back_current])) {
-        //NOTE: This has been double checked with Game Maker 8.1 to work exactly the same, the background_x/y is modified just as object locals are
-        //and also just as one would assume the system to work.
-        //TODO: This should probably be moved to room system.
-        background_x[back_current] += background_hspeed[back_current];
-        background_y[back_current] += background_vspeed[back_current];
-        if (background_htiled[back_current] || background_vtiled[back_current]) {
-          draw_background_tiled_ext(background_index[back_current], background_x[back_current], background_y[back_current], background_xscale[back_current],
-            background_xscale[back_current], background_coloring[back_current], background_alpha[back_current], background_htiled[back_current], background_vtiled[back_current]);
-        } else {
-          draw_background_ext(background_index[back_current], background_x[back_current], background_y[back_current], background_xscale[back_current], background_xscale[back_current], 0, background_coloring[back_current], background_alpha[back_current]);
-        }
-      }
-    }
-  }
-}
-
-static inline void draw_insts()
-{
-  // Apply and clear stored depth changes.
-  for (map<int,pair<double,double> >::iterator it = id_to_currentnextdepth.begin(); it != id_to_currentnextdepth.end(); it++)
-  {
-    enigma::object_graphics* inst_depth = (enigma::object_graphics*)enigma::fetch_instance_by_id((*it).first);
-    if (inst_depth != NULL) {
-      drawing_depths[(*it).second.first].draw_events->unlink(inst_depth->depth.myiter);
-      inst_iter* mynewiter = drawing_depths[(*it).second.second].draw_events->add_inst(inst_depth->depth.myiter->inst);
-      if (instance_event_iterator == inst_depth->depth.myiter) {
-        instance_event_iterator = inst_depth->depth.myiter->prev;
-      }
-      inst_depth->depth.myiter = mynewiter;
-    }
-  }
-  id_to_currentnextdepth.clear();
-
-  if (enigma::particles_impl != NULL) {
-    const double high = numeric_limits<double>::max();
-    const double low = drawing_depths.rbegin() != drawing_depths.rend() ? drawing_depths.rbegin()->first : -numeric_limits<double>::max();
-    (enigma::particles_impl->draw_particlesystems)(high, low);
-  }
-
-  d3dmgr->EndShapesBatching(); //This is needed bacause tiles are individual models
-}
-
-/**
-  Handles tile drawing for a view; returns whether to break the view loop.
-  @return Returns 0 if all is well, or non-zero if the view loop should be broken.
-*/
-static inline int draw_tiles()
-{
-  for (enigma::diter dit = drawing_depths.rbegin(); dit != drawing_depths.rend(); dit++)
-  {
-    if (dit->second.tiles.size())
-    {
-      for (unsigned int t = 0; t<drawing_depths[dit->second.tiles[0].depth].tilevector.size(); ++t){
-        enigma_user::texture_set(drawing_depths[dit->second.tiles[0].depth].tilevector[t][0]);
-        //d3d_model_part_draw(drawing_depths[dit->second.tiles[0].depth].tilelist, drawing_depths[dit->second.tiles[0].depth].tilevector[t][1], drawing_depths[dit->second.tiles[0].depth].tilevector[t][2]);
-      }
-    }
-    enigma::inst_iter* push_it = enigma::instance_event_iterator;
-    //loop instances
-    for (enigma::instance_event_iterator = dit->second.draw_events->next; enigma::instance_event_iterator != NULL; enigma::instance_event_iterator = enigma::instance_event_iterator->next) {
-      enigma::object_graphics* inst = ((object_graphics*)enigma::instance_event_iterator->inst);
-      if (inst->myevent_draw_subcheck())
-        inst->myevent_draw();
-      if (enigma::room_switching_id != -1)
-        return 1;
-    }
-    enigma::instance_event_iterator = push_it;
-    //particles
-    if (enigma::particles_impl != NULL) {
-      const double high = dit->first;
-      dit++;
-      const double low = dit != drawing_depths.rend() ? dit->first : -numeric_limits<double>::max();
-      dit--;
-      (enigma::particles_impl->draw_particlesystems)(high, low);
-    }
-  }
-  return 0;
-}
-
-void clear_view(float x, float y, float w, float h, float angle, bool showcolor)
-{
-	d3d_set_projection_ortho(x, y, w, h, angle);
-
-  DWORD clearflags = 0;
-  int clearcolor = 0;
-	if (background_showcolor)
-	{
-		clearcolor = ((int)background_color) & 0x00FFFFFF;
-		// clear the window to the background color
-    clearflags |= D3DCLEAR_TARGET;
-	}
-
-	// Clear the depth buffer if 3d mode is on at the beginning of the draw step.
-	if (enigma::d3dMode)
-		clearflags |= D3DCLEAR_ZBUFFER;
-
-  d3dmgr->Clear(0, NULL, clearflags, D3DCOLOR_XRGB(COL_GET_R(clearcolor), COL_GET_G(clearcolor), COL_GET_B(clearcolor)), 1.0f, 0);
-}
-
-static inline void draw_gui()
-{
-  int culling = d3d_get_culling();
-  bool hidden = d3d_get_hidden();
-  d3d_set_culling(rs_none);
-  d3d_set_hidden(false);
-
-  bool stop_loop = false;
-  for (enigma::diter dit = drawing_depths.rbegin(); dit != drawing_depths.rend(); dit++)
-  {
-    enigma::inst_iter* push_it = enigma::instance_event_iterator;
-    //loop instances
-    for (enigma::instance_event_iterator = dit->second.draw_events->next; enigma::instance_event_iterator != NULL; enigma::instance_event_iterator = enigma::instance_event_iterator->next) {
-      enigma::object_graphics* inst = ((object_graphics*)enigma::instance_event_iterator->inst);
-      if (inst->myevent_drawgui_subcheck())
-        inst->myevent_drawgui();
-      if (enigma::room_switching_id != -1) {
-        stop_loop = true;
-        break;
-      }
-    }
-    enigma::instance_event_iterator = push_it;
-    if (stop_loop) break;
-  }
-
-  // reset the culling
-  d3d_set_culling(culling);
-  // only restore hidden if the user didn't change it
-  if (!d3d_get_hidden()) {
-    d3d_set_hidden(hidden);
-  }
-}
-
 namespace enigma_user
 {
