+++ ENIGMAsystem/SHELL/Graphics_Systems/None/NONEd3d.cpp
@@ -43,8 +43,8 @@ namespace enigma {
 
 namespace enigma_user
 {
-	void d3d_depth_clear(){}
-	void d3d_depth_clear_value(float value){}
+	void d3d_clear_depth(){}
+	void d3d_clear_depth_value(float value){}
 	void d3d_set_software_vertex_processing(bool software){}
 
 	void d3d_start()
