+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL3/GL3d3d.cpp
@@ -72,7 +72,7 @@ GLenum stenciloperators[8] = {
 namespace enigma_user
 {
 
-void d3d_depth_clear_value(float value) {
+void d3d_clear_depth_value(float value) {
   oglmgr->BlendFunc();
   glClearDepthf(value);
 }
